+++ CHANGELOG.md
@@ -19,6 +19,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
  - Volume ramping configuration moved into volume points (step size, interval, safety range)
  - Fixed bug where QSys camera was reporting the wrong number of supported presets
  - Failing gracefully when a Biamp control fails to load
+ - Using UTC for times
 
 ## [11.7.2] - 2020-02-25
 ### Changed