+++ CHANGELOG.md
@@ -6,6 +6,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+## [12.0.0] - 2020-03-20
 ### Added
  - Added context enum to IVolumePoint and implementations
  - Added mute type enum to IVolumePoint and implementations